Allow for using the RSpec helpers in "around" blocks
Until now the given_filesystem initialization happened in a "before"
block. "before" blocks are only run after the "around" blocks, though,
which made helper calls in "around" blocks fail because they were run
before the initialization.

That is now fixed by moving the initialization to an "around" block
which runs at the very beginning of the test lifecycle.